Add files via upload
This TUNA species data was collected from the Bay of Bengal and the Arabian Sea from 2016 to 2019 [ ~ This is a lat, long, productivity, time series data]. The data is already cleaned, gridded, and geo-referenced. The data is ready to be mapped daily, seasonally, or monthly and also modeled against environmental parameters, and machine learning models can easily be applied to learn from this data.